requirements.txt
ads==0.12.3
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.3
boto3==1.18.3
botocore==1.21.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
distlib==0.3.1
Django==3.2
django-crispy-forms==1.11.2
django-environ==0.4.5
django-recaptcha==2.0.6
django-storages==1.11.1
django-taggit==1.2.0
filelock==3.0.12
google-ads==10.0.0
google-api-core==1.26.3
google-auth==1.29.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
grpcio==1.37.0
gunicorn==20.1.0
httpretty==0.8.10
idna==2.10
isort==5.8.0
jmespath==0.10.0
lazy-object-proxy==1.6.0
Markdown==3.2.1
mccabe==0.6.1
mock==4.0.3
oauthlib==3.1.0
packaging==20.9
pep562==1.0
Pillow==8.2.0
proto-plus==1.18.1
protobuf==3.15.8
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.7.4
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.18.0
python-http-client==3.3.2
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
sendgrid==6.7.0
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
toml==0.10.2
urllib3==1.26.4
virtualenv==20.4.4
Werkzeug==1.0.1
wrapt==1.12.1

